requirements.txt
flask
flask-socketio
# eventlet
pandas
scikit-learn
numpy
joblib
# kafka-python
mysql-connector-python
requests
python-dateutil